ssh-add: show time spent on adding keys
renamed function gensspg2->gensshpg2
gensshpg2:
  failsafe for already present key
  try to determine user from config
  allow username to be different for additional key
  print pubkey for key

+++ ssh.bashrc
@@ -12,28 +12,42 @@ export SSH_AGENT_PID=$SSH_AGENT_PID
 EOF
 unset agent
 
-gensspg() {
-  (
-    set -e
-    if [ $# -eq 0 ]; then
-      printf "You need to provide a Host\n"
-      return 1
-    fi
-    printf "For convenience here a list of your secret gpg keys\n\n"
-    gpg --list-secret-keys --with-subkey-fingerprint --list-options no-show-unusable-subkeys
-    printf "Please provide the gpgkey: "
-    read -r GPGKEY
-    #   CONFGS=( ~/.ssh/config ~gssh/*.ssh )
-    GPGDPATH=~/.ssh/gpg.d
-    SSHHOST=$1
-    SSHEVAL=$(ssh -G $SSHHOST)
-    SSHUSER=$(awk '/^user / {print $2}' <<< $SSHEVAL)
-    KEYPATH=$GPGDPATH/$SSHUSER@$SSHHOST.ed25519
-    ssh-keygen -t ed25519 -f $KEYPATH -C $SSHUSER@$SSHHOST-$USER -P ""
-    gpg --batch --encrypt-files --recipient $GPGKEY $KEYPATH
-    rm $KEYPATH
-    ssh-add $KEYPATH.gpg
-  )
+# Create new keys 
+gensshpg2() {
+    (
+        set -e
+        if [ $# -eq 0 ]; then
+            printf "You need to provide a Host\n"
+            return 1
+        fi
+        if [ -z $SSHPG2ENCRYPTKEY ]; then
+            printf "export SSHPG2ENCRYPTKEY to omit this part.\n\n"
+            printf "For convenience here a list of your secret gpg keys\n\n"
+            gpg --list-secret-keys --with-subkey-fingerprint --list-options no-show-unusable-subkeys
+            printf "Please provide the gpgkey: "
+            read -r $SSHPG2ENCRYPTKEY
+        fi
+        GPGDPATH=~/.ssh/gpg.d
+        # Split user@host
+        SSHHOST=${1#*@}
+        SSHUSER=${1%@*}
+        # Test if we have to lookup the User
+        if [[ $SSHHOST = $SSHUSER ]]; then
+            # Lookup User
+            SSHEVAL=$(ssh -G $SSHHOST)
+            SSHUSER=$(awk '/^user / {print $2}' <<< $SSHEVAL)
+        fi
+        KEYPATH=$GPGDPATH/$SSHUSER@$SSHHOST.ed25519
+        if ! [ -r ${KEYPATH}.gpg ]; then
+            ssh-keygen -t ed25519 -f $KEYPATH -C $SSHUSER@$SSHHOST-$USER -P ""
+            gpg --batch --encrypt-files --recipient $SSHPG2ENCRYPTKEY $KEYPATH
+            rm $KEYPATH
+            ssh-add $KEYPATH.gpg
+        else
+            printf "Key $KEYPATH already exists!\n"
+        fi
+            printf "Pubkey for $SSHUSER@$SSHHOST:\n\n\t%s\n\n" "$(cat $KEYPATH.pub)"
+    )
 }
 
-# vim:ts=4 sts=4 sw=4 et ft=sh
+# vim:ts=4 sts=4 sw=4 et ft=bash
